Use new folder structure in core.
