+++ packages/core/src/components/areas/constituencies/constituencies.data.ts
@@ -1,4 +1,4 @@
-import { AreaType } from '@/components/map/Map'
+import { AreaType } from '@/components/map/map.types'
 
 export type ConstituencyType = AreaType<typeof constituencies>
