--- packages/core/src/components/areas/constituencies/index.ts
@@ -1,2 +1,0 @@
-export { default } from '@/components/areas/constituencies/Constituencies'
-export { ConstituencyType } from '@/components/areas/constituencies/data'
