+++ packages/core/src/components/areas/denmark/Denmark.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import type { DenmarkType } from '@/components/areas/denmark/data'
-import { MapProps } from '@/components/map/Map'
-import Denmark from '@/components/areas/denmark/Denmark'
+import type { DenmarkType } from '@/components/areas/denmark/denmark.data'
+import Denmark from '@/components/areas/denmark'
+import { MapProps } from '@/components/map/map.types'
 
 const Wrapper = (props: MapProps<DenmarkType>) => {
   return (
