+++ packages/core/src/components/areas/denmark/denmark.data.ts
@@ -1,4 +1,4 @@
-import { AreaType } from '@/components/map/Map'
+import { AreaType } from '@/components/map/map.types'
 
 export type DenmarkType = AreaType<typeof full>
