--- packages/core/src/components/areas/denmark/index.ts
@@ -1,2 +1,0 @@
-export { default } from '@/components/areas/denmark/Denmark'
-export { DenmarkType } from '@/components/areas/denmark/data'
