--- packages/core/src/components/areas/islands/index.ts
@@ -1,2 +1,0 @@
-export { default } from '@/components/areas/islands/Islands'
-export { IslandType } from '@/components/areas/islands/data'
