--- packages/core/src/components/areas/municipalities/index.ts
@@ -1,2 +1,0 @@
-export { default } from '@/components/areas/municipalities/Municipalities'
-export { MunicipalityType } from '@/components/areas/municipalities/data'
