+++ packages/core/src/components/areas/municipalities/index.tsx
@@ -1,5 +1,9 @@
-import { MunicipalityType, municipalities } from '@/components/areas/municipalities/data'
-import Map, { MapProps } from '@/components/map/Map'
+import {
+  MunicipalityType,
+  municipalities
+} from '@/components/areas/municipalities/municipalities.data'
+import Map from '@/components/map'
+import { MapProps } from '@/components/map/map.types'
 
 /**
  * Component displaying a map of Denmark with all 98 municipalities.
