--- packages/core/src/components/areas/regions/index.ts
@@ -1,2 +1,0 @@
-export { default } from '@/components/areas/regions/Regions'
-export { RegionType } from '@/components/areas/regions/data'
