+++ packages/core/src/components/index.ts
@@ -1,5 +1,11 @@
-export { default as Municipalities, MunicipalityType } from '@/components/areas/municipalities'
-export { default as Constituencies, ConstituencyType } from '@/components/areas/constituencies'
-export { default as Regions, RegionType } from '@/components/areas/regions'
-export { default as Islands, IslandType } from '@/components/areas/islands'
-export { default as Denmark, DenmarkType } from '@/components/areas/denmark'
+export { default as Municipalities } from '@/components/areas/municipalities'
+export { default as Constituencies } from '@/components/areas/constituencies'
+export { default as Regions } from '@/components/areas/regions'
+export { default as Islands } from '@/components/areas/islands'
+export { default as Denmark } from '@/components/areas/denmark'
+
+export type { MunicipalityType } from '@/components/areas/municipalities/municipalities.data'
+export type { ConstituencyType } from '@/components/areas/constituencies/constituencies.data'
+export type { RegionType } from '@/components/areas/regions/regions.data'
+export type { IslandType } from '@/components/areas/islands/islands.data'
+export type { DenmarkType } from '@/components/areas/denmark/denmark.data'
