+++ packages/core/src/components/map/Map.test.tsx
@@ -1,12 +1,12 @@
 import { useCallback, useMemo, useState } from 'react'
 import { fireEvent, render } from '@testing-library/react'
 // the municipalities component is used to test the generic map component
-import Constituencies from '@/components/areas/constituencies/Constituencies'
-import Denmark from '@/components/areas/denmark/Denmark'
-import Islands from '@/components/areas/islands/Islands'
-import Municipalities from '@/components/areas/municipalities/Municipalities'
-import Regions from '@/components/areas/regions/Regions'
-import { MunicipalityType } from '@/components/areas/municipalities'
+import Constituencies from '@/components/areas/constituencies'
+import Denmark from '@/components/areas/denmark'
+import Islands from '@/components/areas/islands'
+import Municipalities from '@/components/areas/municipalities'
+import Regions from '@/components/areas/regions'
+import { MunicipalityType } from '@/components/areas/municipalities/municipalities.data'
 import { test } from '@/utils'
 
 describe('Map', () => {
